重构部分 record controller 代码; delete record 前进行confirm提示

+++ app/assets/javascripts/components/record.js.coffee
@@ -26,12 +26,14 @@
 
   handleDelete: (e) ->
     e.preventDefault()
-    $.ajax
-      method: 'DELETE'
-      url: "/records/#{ @props.record.id }"
-      dataType: 'JSON'
-      success: () =>
-        @props.handleDeleteRecord @props.record
+    sure = confirm("Are you sure!?")
+    if sure
+      $.ajax
+        method: 'DELETE'
+        url: "/records/#{ @props.record.id }"
+        dataType: 'JSON'
+        success: () =>
+            @props.handleDeleteRecord @props.record
 
   recordForm: ->
     React.DOM.tr null,
